Change layout of data passed to template
We pass the following data layout now:

In pseudo-json:

    {
        "versions": [
            {
                "version": "0.1.0"
                "entries": [ ... ]
            }
        ]
    }

The top-level is a map, so we can pass more data as we see fit.
The "versions" key is a _sorted_ array of version objects, each holding
the version string and a list of entries.
These entries are Fragment objects.

The template was adapted to reflect this.
The tests were adapted to reflect this.

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>