Implement metadata-crawling
This patch implements the first draft of metadata-crawling using either
a command or a path to a script.

Because we altered the default config which we also use in tests, we
have to adapt the `new_command_creates_default_header()` test here, so
that the lines written for additional fields fit the markup.

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>