Remove List/Map Fragment data type
This patch removes the fragment data types of list and map.

This is way too hard to handle in the codebase and not required for an
initial prototype/release, so I remove it now.

It was not implemented anyways, and it could always be added back at
some point in the future.

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>