Add header field type "OneOf"
This patch adds a "One of the following" header field value.
The interactive functionality provides a selection list.

The FragmentDataType type had to be altered to contain either a
"Definite" variant, which is nothing else than a name of a type (one of
"bool", "int" and "string" right now) or a "OneOf".
That change had to be made because serde does not allow to mix tagged
and untagged enums when de/serializing, so the solution is to have two
enums here.

A test for the "OneOf" variant deserialization was added.

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>