Fix: Change filename timestamp format
The fragment files need to be stored with a filename that does not
contain a colon (":") character, because cargo cannot package these
files.

Hence, we alter the filename format in this patch to not contain this
character. The subsecond part is added to avoid collisions in
high-volume projects (most likely not gonna happen, but you'll never
know!).

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>
(cherry picked from commit 6280fd9a3a783450a6f3f601ded6fa947b58d746)
Signed-off-by: Matthias Beyer <mail@beyermatthias.de>